code refactory

+++ GenErrorDiagnostic/src/constraint/graph/Node.java
@@ -47,7 +47,7 @@ abstract public class Node implements Comparable<Node>{
         
     abstract public void incSuccCounter ();
     abstract public void incNestedCounter (int i);
-    abstract public double getSuccCounter ();
+    abstract public int getSuccCounter ();
     abstract public Element getElement ();
     abstract public String getName();
 }
